Only rebuild source when not up-to-date
This will make deploying to multiple environments in a row faster

+++ plans/scm-based.js
@@ -5,6 +5,8 @@ var createDeployment = require('./actions/create-deployment');
 var copyConfig = require('./actions/copy-config');
 var createSymlink = require('./actions/create-symlink');
 var cleanupOldReleases = require('./actions/cleanup-old-releases');
+var fs = require('fs');
+
 // var hipchat = require('../util/hipchat-notification');
 
 module.exports = function(plan, config) {
@@ -17,27 +19,43 @@ module.exports = function(plan, config) {
   });
 
   plan.local('deploy-scm', function (local) {
+    var checkoutProject = function () {
+      local.log('Checking out source code with branch [master]...');
+      local.exec('mkdir -p ' + config.tmp);
+
+      if (argv.branch) {
+        local.exec('git clone ' + source + ' ' + config.tmp, {silent: true});
+        local.exec('cd ' + config.tmp + ' && git checkout ' + argv.branch);
+      } else {
+        local.exec('git clone --depth=1 ' + source + ' ' + config.tmp, {silent: true});
+      }
+
+      local.log('Installing dependencies...');
+      local.exec('cd ' + config.tmp + ' && npm i', {silent: true});
+    };
+
     // var chatNotification = new hipchat(config, local);
     // chatNotification.start();
 
     var source = config.source.replace('{scmuser}', config.scmuser);
 
-    config.tmp = 'tmp/' + (new Date().getTime());
+    config.tmp = 'tmp/deploy'; //' + (new Date().getTime());
 
-    local.log('Checking out source code with branch [master]...');
-    local.exec('mkdir -p ' + config.tmp);
+    // see if there's a previous build. If not, see if it's current.
+    try {
+      var stats = fs.statSync(config.tmp);
+      var localRev = local.exec('cd ' + config.tmp + ' && git rev-parse HEAD', {silent: true}).stdout;
+      var remoteRev = local.exec('cd ' + config.tmp + ' && git rev-parse origin/master', {silent: true}).stdout;
 
-    if (argv.branch) {
-      local.exec('git clone ' + source + ' ' + config.tmp, {silent: true});
-      local.exec('cd ' + config.tmp + ' && git checkout ' + argv.branch);
-    } else {
-      local.exec('git clone --depth=1 ' + source + ' ' + config.tmp, {silent: true});
+      if (localRev !== remoteRev) {
+        local.exec('rm -rf ' + config.tmp);
+        checkoutProject();
+      }
+    } catch (e) {
+      checkoutProject();
     }
 
     copyConfig(config, local);
-
-    local.log('Installing dependencies...');
-    local.exec('cd ' + config.tmp + ' && npm i', {silent: true});
 
     var buildCmd = plan.runtime.options.buildCmd;
 
@@ -66,8 +84,11 @@ module.exports = function(plan, config) {
   });
 
   plan.local('deploy-scm', function (local) {
-    local.log('Cleaning up tmp/');
-    local.exec('rm -rf ' + config.tmp);
+    // only cleanup if we're deploying a non-master branch
+    if (argv.branch) {
+      local.log('Cleaning up tmp/');
+      local.exec('rm -rf ' + config.tmp);
+    }
 
     // var chatNotification = new hipchat(config, local);
     // chatNotification.finish();
